refactor(api): split logout from user service
